config: refactor config to be its own directory
Moving Config which are public into one file and separating out the
things concerning the file handling and parsing and making them
pub(crate)
