+++ src/main.rs
@@ -23,11 +23,13 @@ mod platforms;
 mod system_io;
 
 use crate::comm::dbus::interfaces::{ConfigureInterface, ControlInterface, StatusInterface};
+use crate::config::system_config::system_config;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error>> {
     env_logger::init();
-
+    // call to initialise
+    let _ = system_config();
     // Upon load, the daemon will search each fpga device and determine what platform it is
     // based on its name in /sys/class/fpga_manager/{device}/name
     let status_interface = StatusInterface {};
